Fixed the printing with hex ASCII

+++ tokens.hpp
@@ -34,6 +34,6 @@
   };
   extern int yylineno;
   extern char* yytext;
-  extern size_t yyleng;
+  extern int yyleng;
   extern int yylex();
 #endif /* TOKENS_HPP_ */
